Fix fail to load image when clicking preview
The original implementation embedded the meta data (height, width,
thumbnail..etc) to url query params for chatkit to load the thumbnail
in chatview. This changed the image url and caused `SignatureDoesNotMatch`
when using s3 asset provider. This fix override the onBind of
ImageMessageView to use another load image function of image loader, so
that we don't need to alter the url.